# perf: enhance virtual Tree perf