### 对比

```tsx
<Tree defaultExpandAll data={100w Data} />
```

**Scroll**: ~400ms ->  ~50ms ⬇️
**Expand**: ~200ms -> ~20ms ⬇️

### 优化

* `onMotionEnd`: 在滚动时，会触发 `onVisibleChange` 事件来检查当前的滚动节点里包含不包含动画节点。如果动画节点被移出屏幕外，则直接触发 `onMotionEnd` 事件（如果不触发就会因为节点被删除而永远卡在缩放动画阶段）。原代码实现逻辑为 `所有节点` - `当前展示节点` 得到 `不可见节点`，然后判断一下 `不可见节点` 是否存在动画节点。这里直接改成在 `当前展示节点` 里检查是否有动画节点，没有就触发 `onMotionEnd` 事件。虽然这会反复触发事件，但是对于 Tree 而言，这个状态变化是原子的。相对安全。
* `defaultExpandAll`: 在收缩时会通过 `expandedKeys` 来计算出需要交由虚拟滚动的 `flattenData`，当设置 `defaultExpandAll` 时，会导致 `expandedKeys` 被塞入所有节点的 `keys`。因而在 diff 计算时相当于遍历了所有节点导致卡顿。而对于子节点，其实不需要将其 `key` 参与计算，因而调整 `defaultExpandAll` 逻辑，仅统计有 `children` 的节点的 `key`。由于这是个 break change（相对可控），需要发个 minor 版本。

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- **新特性**
	- 改进了 `NodeList` 组件的可见性变化处理，提高了性能。
	- 优化了 `Tree` 组件的默认展开逻辑，仅展开具有子节点的项。

- **测试**
	- 更新了测试文件中的 `Tree` 组件属性，改为使用 `defaultExpandedKeys`，以更精确地控制树的初始状态。
	- 修改了 `TreeDraggable` 组件的测试预期值，以反映更准确的拖放操作结果。

- **配置**
	- 更新了 TypeScript 配置，改进了 JSX 处理和测试文件的编译范围。

<!-- end of auto-generated comment: release notes by coderabbit.ai -->